swarm: split into Sans-I/O core + std driver; fix no_std builds
Breaking refactor that brings Swarm in line with the project's Sans-I/O
+ no_std non-negotiable. The user-facing DX is preserved.

Architecture now:

  packages/swarm
    src/events.rs   -- SwarmEvent, SwarmAction, SwarmError,
                       OpenStreamToken (no_std + alloc)
    src/core.rs     -- SwarmCore, the Sans-I/O state machine
                       (no_std + alloc). Consumes TransportEvent plus a
                       caller-supplied now_ms; emits SwarmAction commands
                       for a driver to execute and SwarmEvent
                       notifications for the application. Owns no
                       Transport, reads no clock.
    src/driver.rs   -- Swarm<T: Transport>, thin std wrapper that owns a
                       transport and an Instant clock. Auto-allocates
                       ConnectionIds, translates SwarmAction -> transport
                       calls, synchronously recovers StreamIds for
                       open_user_stream via the same flush loop that
                       executes OpenStream actions.
    src/builder.rs  -- SwarmBuilder unchanged externally; still returns
                       the std driver.
    src/lib.rs      -- crate root. #![cfg_attr(not(feature = "std"),
                       no_std)]. Re-exports SwarmCore and the driver.

Cargo.toml: default feature "std"; no_std build drops getrandom and the
driver module. Core uses alloc::collections::BTreeMap everywhere (no
HashMap).

Sans-I/O contract:
- Application drives the core via:
    on_transport_event(event, now_ms)
    on_tick(now_ms)
    on_stream_opened(conn, stream, token, now_ms)
    on_open_stream_failed(token, reason, now_ms)
- Core emits:
    take_actions() -> Vec<SwarmAction>
    poll_events() -> Vec<SwarmEvent>
- OpenStream actions carry an OpenStreamToken so the driver can report
  the allocated stream id back; the core tracks in-flight opens in
  pending_opens until the driver acknowledges.

Latent no_std bugs fixed in the process:
- packages/multistream-select: added `use alloc::vec;`. The crate was
  declared no_std but the `vec!` macro is not in the prelude under
  no_std; `cargo check --no-default-features` was failing silently
  because no workspace consumer exercised that build.
- packages/ping: added `use alloc::vec;` and `use alloc::format;` for
  the same reason.

Both crates now cleanly build with `--no-default-features` along with
minip2p-swarm itself.

DX preserved on the std driver:
- Swarm::dial(&peer_addr) -> ConnectionId (auto-allocated)
- Swarm::ping(&peer_id) (fires on the first negotiated stream, buffering
  the payload if negotiation is still in progress)
- Swarm::open_user_stream(peer, proto) -> StreamId (synchronous)
- Swarm::send_user_stream / close_user_stream_write / reset_user_stream
- Swarm::poll() -> Vec<SwarmEvent> (no now_ms arg; internal clock)
- SwarmBuilder::new(&keypair)....build(transport)

plan.md: updated the crate taxonomy to list packages/swarm under the
Sans-I/O core (the SwarmCore half) plus a note about the std driver;
Milestone 4 marked DONE including the Sans-I/O split; pending
architectural work section reduced to just mutual TLS.

holepunch-plan.md: standalone plan file for the CLI/hole-punch demo
that will consume this refactored swarm.

Tests: 151 passing (unchanged from before the refactor), zero warnings
across --workspace --all-targets, `cargo check --no-default-features`
succeeds for minip2p-swarm, minip2p-multistream-select, minip2p-ping,
minip2p-identify, minip2p-relay, and minip2p-dcutr.
