+++ packages/swarm/Cargo.toml
@@ -3,15 +3,29 @@ name = "minip2p-swarm"
 version = "0.1.0"
 edition = "2024"
 
+[features]
+default = ["std"]
+std = [
+    "minip2p-core/std",
+    "minip2p-identify/std",
+    "minip2p-identity/std",
+    "minip2p-multistream-select/std",
+    "minip2p-ping/std",
+    "minip2p-transport/std",
+    "thiserror/std",
+    "dep:getrandom",
+]
+
 [dependencies]
-minip2p-core = { path = "../core" }
-minip2p-dcutr = { path = "../dcutr" }
-minip2p-identify = { path = "../identify" }
-minip2p-identity = { path = "../identity" }
-minip2p-multistream-select = { path = "../multistream-select" }
-minip2p-ping = { path = "../ping" }
+minip2p-core = { path = "../core", default-features = false }
+minip2p-identify = { path = "../identify", default-features = false }
+minip2p-identity = { path = "../identity", default-features = false, features = ["ed25519"] }
+minip2p-multistream-select = { path = "../multistream-select", default-features = false }
+minip2p-ping = { path = "../ping", default-features = false }
+minip2p-transport = { path = "../transport", default-features = false }
+thiserror = { version = "2", default-features = false }
+getrandom = { version = "0.4", optional = true }
+
+[dev-dependencies]
 minip2p-relay = { path = "../relay" }
-minip2p-transport = { path = "../transport" }
-getrandom = "0.4"
-
-
+minip2p-dcutr = { path = "../dcutr" }
